+++ lightning_examples/distributed-training/scripts/tutorial_model.py
@@ -1,6 +1,6 @@
 import torch
 import torch.nn.functional as F
-from pytorch_lightning import LightningDataModule, LightningModule, seed_everything, Trainer
+from pytorch_lightning import LightningDataModule, LightningModule
 from torch.utils.data import DataLoader, random_split
 from torchmetrics import Accuracy
 from torchvision import transforms
@@ -81,26 +81,3 @@ class TutorialModule(LightningModule):
 
     def configure_optimizers(self):
         return torch.optim.Adam(self.parameters(), lr=self.hparams.learning_rate)
-
-
-class DDPDataDemoModule(TutorialModule):
-
-    def training_epoch_end(self, outputs):
-        process_id = self.global_rank
-        print(f"{process_id=} saw {len(outputs)} samples total")
-
-    def on_train_end(self):
-        print(f"training set contains {len(self.trainer.datamodule.mnist_train)} samples")
-
-
-if __name__ == "__main__":
-    seed_everything(1)
-    model = DDPDataDemoModule()
-    datamodule = MNISTDataModule(batch_size=1)
-
-    trainer = Trainer(
-        gpus=3,
-        accelerator="ddp",
-        max_epochs=1,
-    )
-    trainer.fit(model, datamodule=datamodule)
